+++ src/gwt/src/org/rstudio/studio/client/server/remote/RemoteServer.java
@@ -2359,6 +2359,28 @@ public class RemoteServer implements Server
       params.set(11, JSONBoolean.getInstance(retryWrite));
       sendRequest(RPC_SCOPE, SAVE_DOCUMENT_DIFF, params, requestCallback);
    }
+   
+   public void formatDocument(String id,
+                              String path,
+                              ServerRequestCallback<SourceDocument> requestCallback)
+   {
+      JSONArray params = new JSONArrayBuilder()
+            .add(id)
+            .add(path)
+            .get();
+      
+      sendRequest(RPC_SCOPE, FORMAT_DOCUMENT, params, requestCallback);
+   }
+   
+   public void formatCode(String code,
+                          ServerRequestCallback<String> requestCallback)
+   {
+      JSONArray params = new JSONArrayBuilder()
+            .add(code)
+            .get();
+      
+      sendRequest(RPC_SCOPE, FORMAT_CODE, params, requestCallback);
+   }
 
    public void checkForExternalEdit(
          String id,
@@ -6895,6 +6917,8 @@ public class RemoteServer implements Server
    private static final String OPEN_DOCUMENT = "open_document";
    private static final String SAVE_DOCUMENT = "save_document";
    private static final String SAVE_DOCUMENT_DIFF = "save_document_diff";
+   private static final String FORMAT_DOCUMENT = "format_document";
+   private static final String FORMAT_CODE = "format_code";
    private static final String CHECK_FOR_EXTERNAL_EDIT = "check_for_external_edit";
    private static final String IGNORE_EXTERNAL_EDIT = "ignore_external_edit";
    private static final String CLOSE_DOCUMENT = "close_document";