+++ src/gwt/src/org/rstudio/studio/client/workbench/commands/Commands.java
@@ -29,6 +29,7 @@ public abstract class
 
    // Source
    public abstract AppCommand reformatCode();
+   public abstract AppCommand reformatDocument();
    public abstract AppCommand newSourceDoc();
    public abstract AppCommand newRNotebook();
    public abstract AppCommand newTextDoc();